requirements.txt
asyncio
aiofiles
aiohttp
beautifulsoup4
img2pdf
natsort
Pillow
PyYAML
python-magic
glob2
pathlib
yaml
